requirements.txt
# GoalFlow Project Dependencies

# Core
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
pyyaml>=6.0
numpy>=1.24.0

# Data processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
tensorboard>=2.14.0

# NuScenes (for real data)
# nuscenes-devkit>=1.1.0  # Uncomment when using real data

# Development
black>=23.0.0
isort>=5.12.0
pytest>=7.4.0
